Add navigation and resources fragments + Finish UserLogin and UserRegister
