+++ src/main/java/at/refugeesCode/kitchencheffrontend/configuration/WebMvcConfiguration.java
@@ -9,7 +9,6 @@ public class WebMvcConfiguration implements WebMvcConfigurer {
 
     @Override
     public void addViewControllers(ViewControllerRegistry registry) {
-        registry.addViewController("/login").setViewName("login");
         registry.addViewController("/").setViewName("index");
         registry.addViewController("/add-meal").setViewName("addMeal");
     }
