+++ src/main/resources/templates/fragments/navigation.html
@@ -1,0 +1,89 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+
+<head></head>
+
+<body>
+
+<div th:fragment="menu">
+    <nav class="navbar navbar-expand-sm bg-dark navbar-dark">
+
+        <a class="navbar-brand" href="#">Kitchen Chef</a>
+
+        <!-- Collapsible Navbar -->
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#collapsibleNavbar">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="collapsibleNavbar">
+
+            <!-- Links -->
+            <ul class="navbar-nav">
+                <li class="nav-item active">
+                    <a class="nav-link" href="index.html" th:href="@{/}"> Home</a>
+                </li>
+                <li class="nav-item margin-left10px">
+                    <div class="dropdown">
+                        <button class="btn color-white bg-dark  dropdown-toggle" type="button" data-toggle="dropdown">
+                            Register
+                            <span class="caret"></span>&nbsp;
+                        </button>
+                        <ul class="dropdown-menu dropdown-menu-custom">
+                            <li>
+                                <form id="register" th:action="@{/register}" th:object="${newUser}" method="post">
+                                    <div class="input-group input-group-custom">
+                                        <input id="username1" type="text" th:field="*{username}" class="form-control"
+                                               name="userName"
+                                               placeholder="Username" required/>
+                                    </div>
+                                    <div class="input-group input-group-custom">
+                                        <input id="password1" type="password" th:field="*{password}"
+                                               class="form-control" name="password"
+                                               placeholder="Password" required>
+                                    </div>
+                                    <!--<div class="input-group input-group-custom">-->
+                                    <!--<input id="conPassword1" type="password" class="form-control" name="conPassword1"-->
+                                    <!--placeholder="Confirm Password" required>-->
+                                    <!--</div>-->
+                                    <div>
+                                        <button type="submit" id="registration" value="Create" class="btn btn-primary">
+                                            Submit
+                                        </button>
+                                    </div>
+                                </form>
+                            </li>
+                        </ul>
+                    </div>
+                </li>
+                <li class="nav-item margin-left10px">
+                    <div class="dropdown">
+                        <button class="btn color-white bg-dark  dropdown-toggle" type="button" data-toggle="dropdown">
+                            Log
+                            In
+                            <span class="caret"></span></button>
+                        <ul class="dropdown-menu dropdown-menu-custom">
+                            <li>
+                                <form th:action="@{/login}" method="post">
+                                    <div class="input-group input-group-custom">
+                                        <input type="text" id="username" name="username" autofocus="autofocus"
+                                               class="form-control"
+                                               placeholder="Username" required>
+                                    </div>
+                                    <div class="input-group input-group-custom">
+                                        <input id="password" type="password" class="form-control" name="password"
+                                               placeholder="Password" required>
+                                    </div>
+                                    <div>
+                                        <button id="login" class="btn btn-primary" value="Login">Submit</button>
+                                    </div>
+                                </form>
+                            </li>
+                        </ul>
+                    </div>
+                </li>
+            </ul>
+        </div>
+    </nav>
+</div>
+
+</body>
+</html>
